# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthtech Scheduling API",
        "description": "Agendamento de consultas: pacientes, profissionais e ciclo de vida da consulta.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor atual"
        }
    ],
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Lista consultas (filtrável por status, profissional, paciente e período)",
                "operationId": "bb7eecafa9f5d72b005274a851db91ea",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "professional_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de consultas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Agenda uma nova consulta",
                "operationId": "ad67700abe95e180abd8d79890eaa7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "professional_id",
                                    "start_at",
                                    "end_at"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer"
                                    },
                                    "professional_id": {
                                        "type": "integer"
                                    },
                                    "start_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta agendada"
                    },
                    "422": {
                        "description": "Validação ou fora da janela de disponibilidade"
                    },
                    "409": {
                        "description": "Conflito de agenda"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Detalhe de uma consulta",
                "operationId": "32f042ec6438bcde1acd9a947431abce",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}/confirm": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Confirma uma consulta agendada",
                "operationId": "41f4ca45b1a35422e429b819d525f56a",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta confirmada"
                    },
                    "422": {
                        "description": "Transição inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}/start": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Inicia o atendimento (Confirmada -> Em Atendimento)",
                "operationId": "235648592291e4da7ba03e7305114b9d",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atendimento iniciado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}/complete": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Conclui o atendimento",
                "operationId": "777724cda30d12bb8c1f7d7a311cbc39",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta concluída"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}/no-show": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Marca não comparecimento",
                "operationId": "35a1172a9c7670ac97aa327fb37587d0",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marcado como não comparecido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}/cancel": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancela uma consulta",
                "operationId": "251d27aea150f6bb973b12c8593d6b70",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta cancelada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}/reschedule": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Remarca uma consulta confirmada (cria uma nova vinculada)",
                "operationId": "249625c670f640cd5d230ea1adfb9f65",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_at",
                                    "end_at"
                                ],
                                "properties": {
                                    "start_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nova consulta criada, original marcada como remarcada"
                    },
                    "422": {
                        "description": "Transição inválida ou fora da disponibilidade"
                    },
                    "409": {
                        "description": "Conflito no novo horário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /auth/login. Enviar como \"Authorization: Bearer {token}\".",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}